requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssselect2==0.7.0
Django==4.1.2
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-cryptography==1.1
django-taggit==3.0.0
django-taggit-serializer==0.1.7
django-tinymce==3.5.0
djangorestframework==3.14.0
frozenlist==1.3.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
lxml==4.9.1
multidict==6.0.2
oscrypto==1.3.0
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
PyPDF2==1.26.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
slackclient==2.9.4
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.1.1
tqdm==4.64.1
tzdata==2022.4
tzlocal==4.2
Unidecode==1.3.6
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
yarl==1.8.1
